Refactor Vote component for improved navigation and error handling
