--- src/App.css
@@ -1,124 +1,0 @@
-/* Container styles */
-.waldorf-vote-landing-page {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  max-width: 1200px;
-  margin: auto;
-}
-
-/* Header styles */
-.waldorf-header {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.waldorf-logo-container {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 10px;
-}
-
-.waldorf-logo {
-  max-width: 80px;
-}
-
-.waldorf-main-content {
-  width: 100%;
-}
-
-/* Description styles */
-.waldorf-description {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 10px;
-  background: rgba(255, 255, 255, 0.9);
-  border-radius: 5px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  color: #555;
-  font-size: 1em;
-  line-height: 1.5;
-}
-
-/* Section styles */
-.waldorf-section {
-  background: #fff;
-  border-radius: 5px;
-  padding: 15px;
-  margin-bottom: 20px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.waldorf-section-title {
-  margin-bottom: 10px;
-  font-size: 1.4em;
-  color: #f89e23;
-  text-align: center;
-}
-
-.waldorf-vote-list {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.waldorf-vote-card {
-  padding: 10px;
-  border-left: 3px solid #f89e23;
-}
-
-.waldorf-link {
-  color: #646cff;
-  text-decoration: none;
-  font-weight: bold;
-}
-
-.waldorf-link:hover {
-  text-decoration: underline;
-}
-
-.waldorf-no-votes {
-  text-align: center;
-  color: #888;
-}
-
-/* Footer styles */
-.waldorf-footer {
-  margin-top: 20px;
-  text-align: center;
-}
-
-.waldorf-credits {
-  color: #888;
-}
-
-.waldorf-credits a {
-  color: #646cff;
-  text-decoration: none;
-}
-
-.waldorf-credits a:hover {
-  text-decoration: underline;
-}
-
-/* Responsive design */
-@media (max-width: 768px) {
-  .waldorf-section-title {
-    font-size: 1.2em;
-  }
-}
-
-@media (max-width: 480px) {
-  .waldorf-section {
-    padding: 10px;
-    margin-bottom: 15px;
-  }
-
-  .waldorf-section-title {
-    font-size: 1em;
-  }
-
-  .waldorf-vote-card {
-    padding: 8px;
-  }
-}
