--- src/admin/utils.js
@@ -1,33 +1,0 @@
-export function generateRandomHash(length = 4) {
-  let hash = "";
-  const characters =
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-
-  for (let i = 0; i < length; i++) {
-    hash += characters.charAt(Math.floor(Math.random() * characters.length));
-  }
-
-  return hash;
-}
-
-export const deepEqual = (a, b) => {
-  if (a === b) return true;
-  if (
-    typeof a !== "object" ||
-    a === null ||
-    typeof b !== "object" ||
-    b === null
-  )
-    return false;
-
-  const keysA = Object.keys(a);
-  const keysB = Object.keys(b);
-
-  if (keysA.length !== keysB.length) return false;
-
-  for (let key of keysA) {
-    if (!keysB.includes(key) || !deepEqual(a[key], b[key])) return false;
-  }
-
-  return true;
-};
