--- src/index.css
@@ -1,135 +1,0 @@
-:root {
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-a:hover {
-  color: #535bf2;
-}
-
-body {
-  margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-}
-
-h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-.button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  transition: border-color 0.25s;
-  cursor: pointer;
-  text-align: center;
-  color: white;
-}
-.button:hover {
-  border-color: #646cff;
-  color: white;
-}
-.button:focus,
-.button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-  color: white;
-}
-
-input {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  transition: border-color 0.25s;
-  cursor: pointer;
-  text-align: center;
-  color: white;
-}
-input:hover {
-  border-color: #646cff;
-  color: white;
-}
-input:focus,
-.button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-  color: white;
-}
-
-.button.disabled {
-  background-color: #5a5a5a;
-  cursor: default;
-}
-.button.disabled:hover {
-  border-color: #5a5a5a;
-}
-
-.button.active {
-  background-color: #f89e23;
-  cursor: default;
-}
-.button.active:hover {
-  border-color: #f89e23;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  .button {
-    background-color: #f9f9f9;
-    color: gray;
-  }
-  .button:hover {
-    background-color: white;
-    color: black;
-  }
-  .button.disabled {
-    background-color: #5a5a5a;
-    cursor: default;
-    color: black;
-  }
-  .button.disabled:hover {
-    border-color: #5a5a5a;
-  }
-  .button.active {
-    background-color: #f89e23;
-    cursor: default;
-    color: black;
-  }
-  .button.active:hover {
-    border-color: #f89e23;
-  }
-  input.button {
-    color: black;
-  }
-}
