--- src/vote.css
@@ -1,77 +1,0 @@
-.container {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.card {
-  width: 60vw;
-  padding: 50px !important;
-  margin: 50px;
-}
-
-@media (max-width: 768px) {
-  .card {
-    width: 100%;
-    padding: 15px !important;
-    margin: 0;
-    min-height: 100vh;
-  }
-}
-
-.option-card {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  padding: 20px;
-  flex: 1 1 45%;
-}
-
-.option-card.selected {
-  border: 5px solid rgb(var(--mdui-color-primary)) !important;
-}
-
-@media (max-width: 768px) {
-  .option-card {
-    width: 100%;
-    flex: 1 1 100%;
-  }
-}
-
-.flex-row {
-  display: flex;
-  flex-direction: row;
-  gap: 20px;
-}
-
-@media (max-width: 768px) {
-  .flex-row {
-    flex-direction: column;
-    gap: 10px;
-  }
-}
-
-.flex-wrap {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-}
-
-.button-container {
-  display: flex;
-  justify-content: space-between;
-}
-
-@media (max-width: 768px) {
-  .button-container {
-    flex-direction: column;
-    gap: 10px;
-  }
-}
-
-@media (max-width: 768px) {
-  .vote-title {
-    font-size: 30px !important;
-  }
-}
